refactor: UserCommandService 인터페이스 및 구현체 분리 (#13)
